+++ src/common.ts
@@ -1,0 +1,5 @@
+export interface Reading {
+    temperature: number, //kelvin
+    altitude: number, //meter
+    datetime: Date
+}
